Move AgentOperationsManager from managers/ to services/
It depends on the service layer, so it belongs in services/ rather than
managers/ where the dependency direction would be inverted. Also deduped
the MinimalTrans boilerplate in the integration tests.
