+++ lib/galaxy/agents/history.py
@@ -14,7 +14,7 @@ from pydantic_ai import Agent
 from pydantic_ai.tools import RunContext
 
 from galaxy.exceptions import MalformedId
-from galaxy.managers.agent_operations import AgentOperationsManager
+from galaxy.webapps.galaxy.services.agent_operations import AgentOperationsManager
 from .base import (
     AgentType,
     BaseGalaxyAgent,
