+++ lib/galaxy/webapps/galaxy/services/agent_operations.py
